chatend/requirements.txt
django>=4.2
djangorestframework==3.12.4
djangorestframework-simplejwt
